requirements.txt
websocket-client
Pillow
fastapi
uvicorn
pydantic